Change Default Form State to Valid
I was initializing the UI fields for user input to the invalid state so
that the user couldn't accidentally (or intentionally) submit empty or
invalid input when opening the modal, but this causes the modal to be
unusable when the form-validation websocket is not implemented.

In a real application, this is probably what we would want since the
validation socket being unreachable would be a serious problem, but for
this project, we need the workshop participants to be able to use as
much of the site as possible without the websockets working. That way,
they can get an idea of how the realtime behavior of the websocket they
are implementing fits into the application, which is an important part
of backend development. For this reason, I have changed the initial
state of these input fields to valid, so that the forms will still work
without a validation socket. The input will still be validated when it
gets submitted. The user just won't see the error messages in realtime
as they type.