Fix UI Bug when Joining Chatroom
There was a bug in the frontend logic when joining or creating a
chatroom.

If the user was creating or joining a chatroom from the modal with
another chatroom already open in the main interface, then the main chat
interface would not update to show all of the new room's data.

Specifically, the connections and state would correctly update, but the
message history and other UI elements from the previous room would not
clear, so the UI would have a mix of messages from the previous room and
the new one.

This bug was the result of a couple of refactors I did previously. I
originally had the frontend architected similarly to how it is
structured right now, but I refactored it to use more event-driven logic
for detecting when an interface should clear/update. This ended up being
unreliable and the UX was kind of clunky, so I refactored back to
explicit function calls, but I forgot to add the checks for existing UI
when moving back from the event-based checks to the explicit coroutines.

I have re-added the checks to the `connectToRoom` function, and I also
found and removed a redundant call to the `room-data` endpoint in the
`joinRoom` function.