Increase the Character Limit for User and Room Names
While testing some UI fixes, I realized that 16 characters was pretty
limiting for room names. I originally chose this value to make sure
names would fit in the UI layout I had put together, but that's a
problem to solve with CSS not db constraints, so I've increased the
character limit for room and user names from 16 to 32.

The columns still use "NOCASE" collation, so validation based on
queries of existing records and UNIQUE constraints will still work
case-insensitively, but we can have three average-length English words
for a room name now.

I also added a `chat.constants` module that defines a `NAME_LENGTH`
value to avoid potential bugs in validation routines that hardcode a
value instead of using whatever value is used in the db schema. By
referencing this constant in both the model definitions and any
validation logic, we can avoid the potential mismatch and let the db act
as the single source of truth for data integrity.