Add Metadata Description to Webpage
I added a description meta tag to the index.html file so that linking to
a page will display the rich content overview. The description is
generic, but its only purpose is to make that content overview appear in
chat apps and whatnot, so it's fine to leave as a generic description
for this workshop project.

This should have been in there from the start, but I forgot about it
while working on Javascript stuff.