Add Missing Test Case for Chat History WebSocket
I forgot to add a test case to check that the messages sent by the chat
history socket are older than the reference timestamp provided by the
client, so I have added a test case to the
`tests/websockets/chat_history_test.py` file to cover this part of the
requirements.

I also changed a stupid query in the `generate_message_history` fixture
that called `get_user_by_id` and then directly assigned the id of the
result to a variable. This is dumb since it's just a more complicated
way to hard-code the ID. I changed this to lookup the user by name so
that we avoid potential breakage due to hard-coded PKs not matching up.