Change Requirements for Chat History WebSocket
I decided to change the requirements for the chat-history websocket to
include an ordering argument/field. The websocket route now needs to
support returning results that are either older than the reference
timestamp in descending order (the same as before) or newer than the
reference timestamp in ascending order.

I decided to make this change since it is more representative of an
actual backend route in a real application. It's very easy to fall into
the trap of letting the frontend UI logic dictate how the backend works,
but it's better to design endpoints that provide some kind of
backend-specific functionality, and then the frontend can use those
endpoints in different ways.

In this case, the purpose of the chat-history websocket is to provide
efficient access to chat history in configurable chunks. The concept of
a chat history can be thought of as a timeline, and a timeline makes
sense to move either forward or backward from an arbitrary reference
point. The previous requirements only supported moving backwards in
time, but for a general chat history endpoint as a backend dev would
likely implement in a real application, being able to move in either
direction should be supported.

Practical examples of things this might be used for are adding a chat log
page that includes paginated messages, or adding a true infinite scroll
where the UI only shows a view of say 10 messages, and as the user
scrolls up or down, the 10 messages on screen dynamically update to show
the older or newer messages in the history. These are just examples of
course, and basing backend design decisions off of possible future UI
features is a bad habit to get into as well. The purpose of allowing
chunking in either direction is simply because that's a characteristic
of a historical timeline, so it should be supported by an endpoint that
is providing that kind of functionality.

I have also added test cases to the
`tests/websockets/chat_history_test.py` file that check these new
requirements.