Fix Bug in Websockets Auth Test Cases
When writing the test cases for ensuring that websockets are
authenticated, I forgot to assert False when no exception is raised.

Because we are using the InvalidStatusCode exception of the websockets
library to assert the status is a 401 on auth failure, the test will
silently pass if no exception is raised. This will cause a false passing
test if no auth checking is performed at all, which would be a major
bug.

I have fixed this by adding an else block that asserts False to the
try-except blocks that are checking the status code.