Implement Password Hashing for User Login Tokens
The login system is very simple for this application. When a new user
is created, We generate a JWT which will be used for session auth
when making async requests from the frontend after a user has logged
in. This JWT is stored in the db, and we run it through a couple of
hash functions to get a random byte string and then that string is sent
back to the frontend after creating the user in order to act as
the user's username + password combo for login purposes.

I originally stored the hashed token in the db to simplify the
process of handling user login since the main purpose of this project
was to practice the websockets implementation. This is not secure
because the generated token is a proxy for the user's password, so we
are storing passwords in plaintext.

I wasn't worried about this initially because this is just a workshop
app intended for practice implementing websockets in Python, but then I
remembered that `left-pad` is a thing. It's entirely possible that some
rando on the internet could find this repo and decide to use it as a
baseline for a production server. That would be bad, so I have added
password hashing to the auth system and removed the plaintext tokens
from the db.

This is a bit more complicated than a typical auth system since we are
still only asking for a single value from the user on login. This means
that we have to somehow derive a value from the user's plaintext
*password* that we can use to locate the hashed password in the db. The
problem is that any derived value needs to be reproducible based on the
plaintext password, which means if it's some kind of hashing function,
it can't be salted, which means that the password could be reconstructed
via a rainbow table or another brute forcing mechanism.

Most sites just look up the hashed password by the username that the
user provides with their login, but we can't do that with this workflow.

I decided to simply append the user_id to the login token. Since we're
using the sha1 hexdigest to shorten the login token after generation,
it will always be 40 characters long, so we can simply split the string
at 40 characters to get the plaintext password and the user_id. Then we
can look up the password hash via the user_id and compare the plaintext
password with the hashed password via argon2.

This works because the user_id is not a sensitive field. It is shared
with other clients on the frontend in order to identify different users
connected to a chatroom, so it isn't used for anything that would be
problematic if an attacker had access to it. This allows us to avoid
storing the plaintext password token in the db without requiring the
user to enter more than one data point at login.

I also ran neovim's autoformat over the pyproject.toml file, so that
file has some formatting changes included in the diff. The only true
change in the pyproject.toml file is the addition of the `argon2-cffi`
dependency.